Implement more MR comments
